+++ include/GenThread.h
@@ -6,7 +6,7 @@
 
 #include <boost/random/mersenne_twister.hpp>
 
-#include "Ini.h"
+#include "Configuration.h"
 
 namespace fs = std::filesystem;
 
@@ -24,7 +24,7 @@ class Generator {
 
     boost::mt19937 randomNumberSequence;
 
-    std::string getMesh(const std::list<GrassIni2::GrassMesh> &meshList, const std::string &cat);
+    std::string getMesh(const std::vector<ObjectPlacementPossibility>& placements, const std::string &cat);
 
     float getRandom(float min, float max);